# server/resources/swagger.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Application Service Center API
  description: Application Service Center API
  externalDocs:
    description: Find out more
    url: https://github.wdf.sap.corp/ApplicationServices/admin_v2
servers:
  - url: /api/v1
paths:
...
  /apps:
    get:
      summary: List all apps
      tags:
        - Apps
      parameters:
        - in: query
          name: category
          schema:
            type: string
          required: false
          description: The category of apps to return
      responses:
        '200':
          description: A array of Apps
          content:
            application/json:    
              schema:
                $ref: "#/components/schemas/Apps"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      summary: Create a App
      tags:
        - Apps
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/App'
      responses:
        '201':
          description: Standard create response
          content:
            application/json:    
              schema:
                $ref: "#/components/schemas/Standard"
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /apps/{app_id}:
    parameters:
      - name: app_id
        in: path
        required: true
        description: The id of the app to retrieve
        schema:
          type: integer
    get:
      summary: Info for a specific app
      tags:
        - Apps
      responses:
        '200':
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/App"
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      summary: Update a specific app
      tags:
        - Apps
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/App'
      responses:
        '200':
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Standard"
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Delete a specific app
      tags:
        - Apps
      responses:
        '200':
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Standard"
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # App Bulk Data
...
  /contacts/{contact_id}/apps:
    parameters:
      - name: contact_id
        in: path
        required: true
        description: The id of the contact
        schema:
          type: integer
    get:
      summary: Get all apps for a specific contact
      tags:
        - Contacts
      responses:
        '200':
          description: A array of apps for a specific contact
          content:
            application/json:    
              schema:
                $ref: "#/components/schemas/Apps"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  # Settings
...
security:
  - ApiKeyAuth: []

# Components/Schemas
components:
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: query
      name: access_token
  schemas:
...
    App:
      type: object
      required:
        - app_name
      properties:
        app_id:
          type: integer
          format: int64
        app_name:
          type: string
        category:
          type: string
        etag:
          type: string
        status:
          type: string
        created:
          type: string
          format: date-time
        modified:
          type: string
          format: date-time
        notes:
          type: string
        monitoring_url:
          type: string
        go_live:
          type: string
          format: date-time
        retired:
          type: string
          format: date-time
      example:
        app_name: 'SAP Relay'
        category: 'iOS'
        status: 'Active'
        notes: 'v2 of our popular messaging app'
        monitoring_url: 'https://'
        go_live: '2019-01-01 00:00:10'
    Apps:
      type: array
      items:
        $ref: "#/components/schemas/App"
...
    Standard:
      type: object
      required:
        - status
        - message
      properties:
        status:
          type: string
        message:
          type: string
    Error:
      type: object
      required:
        - data
      properties:
        data:
          type: string